add function hook
